+++ src/app/page.js
@@ -1,14 +1,16 @@
 import Link from 'next/link';
-import { getPosts } from '../utils/mdx-utils';
+import { getPosts } from '@/utils/mdx-utils';
 
-import Footer from '../components/Footer';
-import Header from '../components/Header';
-import Layout, { GradientBackground } from '../components/Layout';
-import ArrowIcon from '../components/ArrowIcon';
-import { getGlobalData } from '../utils/global-data';
-import SEO from '../components/SEO';
+import Footer from '@/components/Footer';
+import Header from '@/components/Header';
+import Layout, { GradientBackground } from '@/components/Layout';
+import ArrowIcon from '@/components/ArrowIcon';
+import { getGlobalData } from '@/utils/global-data';
+import SEO from '@/components/SEO';
 
-export default function Index({ posts, globalData }) {
+export default function Index() {
+  const posts = getPosts();
+  const globalData = getGlobalData();
   return (
     <Layout>
       <SEO title={globalData.name} description={globalData.blogTitle} />
@@ -67,9 +69,3 @@ export default function Index({ posts, globalData }) {
   );
 }
 
-export function getStaticProps() {
-  const posts = getPosts();
-  const globalData = getGlobalData();
-
-  return { props: { posts, globalData } };
-}
