+++ src/components/Footer.js
@@ -1,3 +1,4 @@
+'use client';
 const sunIcon = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
